+++ packages/mint-components/src/components/tax-and-cash/TaxForm.stories.tsx
@@ -25,6 +25,7 @@ const stepOneProps: StoryDemoData<UseUserInfoFormResult> = {
   states: {
     step: "1",
     hideSteps: false,
+    hideState: false,
     loading: false,
     disabled: false,
     isPartner: false,
@@ -34,12 +35,26 @@ const stepOneProps: StoryDemoData<UseUserInfoFormResult> = {
       firstName: "Bob",
       lastName: "Testerson",
       email: "bobtesterson@example.com",
+      phoneNumberCountryCode: "US",
+      phoneNumber: "555 555 5555",
       countryCode: "US",
       currency: "CAD",
+      address: "2/1337 Johnson Street",
+      city: "Vancouver",
+      state: "BC",
+      postalCode: "555 555",
       errors: {},
     },
   },
   data: {
+    regionLabelEnum: "STATE",
+    regions: [],
+    phoneCountries: [
+      {
+        countryCode: "CA",
+        displayName: "Canada",
+      },
+    ],
     countries: [
       {
         countryCode: "CA",
@@ -120,6 +135,7 @@ const stepOneProps: StoryDemoData<UseUserInfoFormResult> = {
   refs: {
     formRef: { current: null },
     currencyRef: { current: null },
+    phoneCountryRef: { current: null },
   },
 };